hey guess what still procrastinating and now I added the ability to move straight up or down in the camera, tbf this is something I wish other engines had (i could be wrong though i feel like you can move like this in unreal and unity but shit idc as far as im concerned i am the inventor of this fuck you)

+++ Enge/Application/Application.h
@@ -34,8 +34,8 @@ namespace Enge {
 
 		bool shouldClose = false;
 
-		float deltaTime = 0.0f;
-		float lastFrame = 0.0f;
+		double deltaTime = 0.0f;
+		double lastFrame = 0.0f;
 	};
 
 }
